fix: remove form methods from empty state
